pyproject.toml
# ==============================================================================
#                              Build System
# ==============================================================================
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# ==============================================================================
#                               Project
# ==============================================================================
[project]
name = "market-beacon"
version = "0.0.0"
description = "A Python bot to retrieve and analyze trade data from the Bitget exchange."
authors = [
    { name = "David Young", email = "davidsamuelyoung@protonmail.com" },
]
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["cryptocurrency", "bot", "trading", "bitget", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "loguru>=0.7.2",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dateutil>=2.8.2"
]

[project.urls]
Homepage = "https://github.com/the-user-created/market-beacon"
Issues = "https://github.com/the-user-created/market-beacon/issues"
Repository = "https://github.com/the-user-created/market-beacon"


# ==============================================================================
#                                 Tools
# ==============================================================================

# ---------------------------- Dependency Groups -------------------------------
[dependency-groups]
dev = [
    "bump-my-version>=1.2.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",
    "pre-commit>=3.6.0",
    "types-requests",
]


# --------------------------------- Ruff ---------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W", "F", # pycodestyle, Pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "N", # pep8-naming
    "PT", # flake8-pytest-style
    "SIM", # flake8-simplify
    "UP", # pyupgrade
    "RUF", # Ruff-specific rules
]

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false

# ------------------------------ Bump My Version -------------------------------
[tool.bumpversion]
commit = true
tag = true
message = "chore(version): Bump version to {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

# ------------------------------ setuptools ------------------------------------
[tool.setuptools.packages.find]
where = ["src"]

# --------------------------------- Codespell ----------------------------------
[tool.codespell]
skip = "*.lock,*.json,*.svg"
ignore-words-list = "sur,alo,larg,gam,hig"
quiet-level = 2
